requirements.txt
streamlit
openai
chromadb
pysqlite3-binary
PyPDF2
protobu==3.20